chore: unify package.json
