+++ app.config.js
@@ -1,0 +1,3 @@
+export default {
+  entryPoint: './example/App',
+};
