+++ example/babel.config.js
@@ -1,3 +1,8 @@
-module.exports = {
-  presets: ['babel-preset-expo'],
+module.exports = function (api) {
+  if (api) {
+    api.cache(true);
+  }
+  return {
+    presets: ['babel-preset-expo'],
+  };
 };
