--- example/index.web.js
@@ -1,9 +1,0 @@
-// Entry point for react-native-web
-import App from './App';
-import {AppRegistry} from 'react-native';
-import {name as appName} from './app.json';
-
-AppRegistry.registerComponent(appName, () => App);
-AppRegistry.runApplication(appName, {
-  rootTag: document.getElementById('root'),
-});
