--- example/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "example",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "android": "react-native run-android",
-    "ios": "react-native run-ios",
-    "start": "react-native start",
-    "web": "expo start --web"
-  },
-  "dependencies": {
-    "@react-native-community/segmented-control": "file:../",
-    "expo": "^37.0.7",
-    "react": "16.11.0",
-    "react-dom": "16.11.0",
-    "react-native": "0.62.2",
-    "react-native-web": "^0.12.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/runtime": "^7.6.2",
-    "babel-preset-expo": "^8.1.0"
-  },
-  "jest": {
-    "preset": "react-native"
-  }
-}
