+++ package.json
@@ -11,8 +11,11 @@
     "react-native-segmented-control.podspec"
   ],
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
+    "start": "react-native start",
     "test": "yarn flow && yarn lint",
+    "ios": "cd example && react-native run-ios",
+    "web": "expo web",
+    "android": "react-native run-android",
     "flow": "flow",
     "lint": "eslint . --cache",
     "tsc": "tsc --noEmit"
@@ -22,23 +25,28 @@
     "react-native": ">=0.57"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.0",
-    "@babel/runtime": "^7.4.2",
-    "@react-native-community/eslint-config": "^1.0.0",
+    "@babel/core": "^7.6.2",
+    "@babel/runtime": "^7.6.2",
+    "@expo/webpack-config": "^0.12.15",
+    "@react-native-community/eslint-config": "^2.0.0",
     "@semantic-release/git": "^8.0.0",
     "@types/react": "^16.9.17",
     "@types/react-native": "^0.62.1",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^25.1.0",
+    "babel-preset-expo": "^8.2.1",
     "eslint": "^6.8.0",
     "eslint-plugin-prettier": "^3.1.2",
+    "expo": "^37.0.8",
     "flow-bin": "^0.113.0",
     "husky": "^4.0.10",
     "jest": "^25.2.4",
     "lint-staged": "^10.0.0",
     "metro-react-native-babel-preset": "0.56.0",
     "react": "16.11.0",
+    "react-dom": "16.11.0",
     "react-native": "^0.62.2",
+    "react-native-web": "^0.11.0",
     "react-test-renderer": "16.11.0",
     "rimraf": "^3.0.2",
     "semantic-release": "^17.0.4",
